[IMP] gpautoservices: validate product analytic assignment
This commit adds a new validation in sale.order lines to prevent adding
a product if the product has requires_analytic_assignment set to True,
and none of the assigned analytic accounts by the distribution model
have has_product_relation enabled.

This ensures that sales order lines with products requiring analytic
assignment are properly linked to analytic accounts associated with an
available machine.

Related to T#94330